update patient register
